Replace Space trait by a Space struct
Merge the four ‘acting’ methods of a Space trait into a single one.
With that, the trait is reduced to only two methods: get_file_suffix
and fill_channels.  At this point, the code can be simplified by using
a Struct field with a file suffix constant and a function pointer which
simplifies the usage and removes the need for vtables or unused &self
method argument.