Support other number of channels than three
With some arguably unsafe code, add support for other number of channels
than just three.  This is done by changing the fill_channels callback
to essentially accept pointer to the pixel in the first channel and
width of the image as arguments.  By adding the width to the pointer,
the function can gain access to corresponding pixel in other channels.