Convert to RGB8 after resizing and cropping
So far, the code would load the image, convert it to RGB8 (if not already
that) and then resize and crop it.  Change the order of operations such
that resizing and cropping happens beforehand.  For greyscale images this
will speed resizing up; for images with 16-bit depth this will likely
improve resizing quality.