Get rid of Confirmer type
Move `confirm` method from the Confirmer type to Opts struct.  This
simplifies the code by removing the need to create another object and
pass it around.  Furthermore, remove the mutex that Confirmer used
and instead take advantage of StdinLock which can be used to
synchronise asking for confirmation between threads.